Implement a source code to calculate average of age based on location's information
Writer: Soyeong Park

This source code was made to understand MapReduce and we implement to
calculate average of age based on location.

So to speak, This program is a source to calculate average of age based
on location’s information after xml is parsed. And I used
CountLocation’s program source code.